# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Talking Shaha Core Garage API
  version: 0.1.0
  description: Core garage, vehicle dashboard, vehicle parts, and timeline (service history) API for the backend flow.
servers:
  - url: /api/v1
paths:
  /vehicles:
    get:
      summary: List demo user's garage vehicles
      tags: [ Vehicles ]
      responses:
        '200':
          description: Vehicles in the current garage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
    post:
      summary: Add vehicle to garage
      tags: [ Vehicles ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'
      responses:
        '201':
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/ValidationError'
  /vehicles/{vehicleId}:
    patch:
      summary: Update vehicle
      tags: [ Vehicles ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'
      responses:
        '200':
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete vehicle with its parts and events
      tags: [ Vehicles ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      responses:
        '204':
          description: Vehicle deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/dashboard:
    get:
      summary: Get vehicle dashboard with parts forecast
      tags: [ Vehicles ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      responses:
        '200':
          description: Vehicle dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDashboard'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/parts:
    get:
      summary: List installed vehicle parts
      tags: [ Parts ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      responses:
        '200':
          description: Installed parts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartList'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Add installed vehicle part
      tags: [ Parts ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartRequest'
      responses:
        '201':
          description: Created part with calculated lifetime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/parts/{partId}:
    patch:
      summary: Update installed vehicle part
      tags: [ Parts ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
        - name: partId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartRequest'
      responses:
        '200':
          description: Updated part with recalculated lifetime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/timeline:
    get:
      summary: List vehicle timeline events
      tags: [ Timeline ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
        - name: type
          in: query
          required: false
          description: Optional filter by event type
          schema:
            $ref: '#/components/schemas/TimelineEventType'
      responses:
        '200':
          description: Timeline events, most recent first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEventList'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/timeline/refuel:
    post:
      summary: Add refuel event
      tags: [ Timeline ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefuelEventRequest'
      responses:
        '201':
          description: Created refuel event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/timeline/trip:
    post:
      summary: Add trip event
      tags: [ Timeline ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripEventRequest'
      responses:
        '201':
          description: Created trip event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/timeline/maintenance:
    post:
      summary: Add maintenance event
      tags: [ Timeline ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceEventRequest'
      responses:
        '201':
          description: Created maintenance event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/timeline/part:
    post:
      summary: Add part replacement event
      tags: [ Timeline ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartEventRequest'
      responses:
        '201':
          description: Created part replacement event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehicles/{vehicleId}/analytics:
    get:
      summary: Get analytics overview
      tags: [ Analytics ]
      parameters:
        - $ref: '#/components/parameters/VehicleId'
        - name: period
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AnalyticsPeriod'
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Resource belongs to another user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [ code, message, fields ]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request contains invalid fields
        fields:
          type: object
          additionalProperties:
            type: string
    CreateVehicleRequest:
      type: object
      required: [ brand, model, productionYear, mileageKm, fuelType ]
      properties:
        brand:
          type: string
          maxLength: 80
        model:
          type: string
          maxLength: 120
        productionYear:
          type: integer
          minimum: 1900
          maximum: 2026
        color:
          type: string
          maxLength: 40
        mileageKm:
          type: integer
          minimum: 0
        fuelType:
          $ref: '#/components/schemas/FuelType'
        engineDescription:
          type: string
          maxLength: 80
        vin:
          type: string
          minLength: 17
          maxLength: 17
        photoUrl:
          type: string
          maxLength: 500
    Vehicle:
      allOf:
        - $ref: '#/components/schemas/CreateVehicleRequest'
        - type: object
          required: [ id ]
          properties:
            id:
              type: string
              format: uuid
    VehicleDashboard:
      type: object
      required: [ vehicle, maintenanceForecast, recentEvents ]
      properties:
        vehicle:
          $ref: '#/components/schemas/Vehicle'
        maintenanceForecast:
          $ref: '#/components/schemas/MaintenanceForecast'
        recentEvents:
          type: array
          items:
            $ref: '#/components/schemas/RecentEvent'
    MaintenanceForecast:
      type: object
      required: [ overallStatus, updatedAt, parts ]
      properties:
        overallStatus:
          $ref: '#/components/schemas/PartStatus'
        nextServiceInKm:
          type: integer
          nullable: true
        updatedAt:
          type: string
          format: date-time
        parts:
          type: array
          items:
            $ref: '#/components/schemas/Part'
    RecentEvent:
      type: object
      description: Compact timeline event shown on the dashboard.
      required: [ id, type, eventDateTime ]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TimelineEventType'
        title:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        eventDateTime:
          type: string
          format: date-time
    PartList:
      type: object
      required: [ parts ]
      properties:
        parts:
          type: array
          items:
            $ref: '#/components/schemas/Part'
    CreatePartRequest:
      type: object
      required: [ name, category, installedAt, installedMileageKm ]
      properties:
        name:
          type: string
          maxLength: 255
        category:
          $ref: '#/components/schemas/PartCategory'
        installedAt:
          type: string
          format: date
        installedMileageKm:
          type: integer
          minimum: 0
        expectedLifetimeKm:
          type: integer
          minimum: 1
          nullable: true
        description:
          type: string
          maxLength: 2000
          nullable: true
        cost:
          type: number
          minimum: 0
          exclusiveMinimum: true
          nullable: true
        photoUrls:
          type: array
          nullable: true
          items:
            type: string
    UpdatePartRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        category:
          $ref: '#/components/schemas/PartCategory'
        installedAt:
          type: string
          format: date
        installedMileageKm:
          type: integer
          minimum: 0
        expectedLifetimeKm:
          type: integer
          minimum: 1
        description:
          type: string
          maxLength: 2000
        cost:
          type: number
          minimum: 0
          exclusiveMinimum: true
        photoUrls:
          type: array
          nullable: true
          items:
            type: string
    Part:
      type: object
      required: [ id, name, category, installedAt, installedMileageKm, remainingKm, remainingPercent, status ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          $ref: '#/components/schemas/PartCategory'
        installedAt:
          type: string
          format: date
        installedMileageKm:
          type: integer
        expectedLifetimeKm:
          type: integer
          nullable: true
        remainingKm:
          type: integer
          nullable: true
        remainingPercent:
          type: integer
          nullable: true
        status:
          $ref: '#/components/schemas/PartStatus'
        description:
          type: string
          nullable: true
        cost:
          type: number
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
    UpdateVehicleRequest:
      type: object
      description: Partial vehicle update; only the provided fields are applied.
      properties:
        brand:
          type: string
          maxLength: 80
        model:
          type: string
          maxLength: 120
        productionYear:
          type: integer
          minimum: 1900
          maximum: 2100
        color:
          type: string
          maxLength: 40
        mileageKm:
          type: integer
          minimum: 0
        fuelType:
          $ref: '#/components/schemas/FuelType'
        engineDescription:
          type: string
          maxLength: 80
        vin:
          type: string
          minLength: 17
          maxLength: 17
        photoUrl:
          type: string
          maxLength: 500
    CreateRefuelEventRequest:
      type: object
      required: [ eventDateTime, mileageKm, liters, cost, fuelType ]
      properties:
        eventDateTime:
          type: string
          format: date-time
        mileageKm:
          type: integer
          minimum: 0
        liters:
          type: number
          minimum: 0
          exclusiveMinimum: true
        cost:
          type: number
          minimum: 0
          exclusiveMinimum: true
        fuelType:
          $ref: '#/components/schemas/FuelType'
        fuelName:
          type: string
          nullable: true
        stationName:
          type: string
          nullable: true
    CreateTripEventRequest:
      type: object
      required: [ eventDateTime, endMileageKm, durationMinutes ]
      properties:
        eventDateTime:
          type: string
          format: date-time
        startMileageKm:
          type: integer
          minimum: 0
          nullable: true
        endMileageKm:
          type: integer
          minimum: 0
        route:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          minimum: 1
    CreateMaintenanceEventRequest:
      type: object
      required: [ eventDateTime, mileageKm, name ]
      properties:
        eventDateTime:
          type: string
          format: date-time
        mileageKm:
          type: integer
          minimum: 0
        name:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        cost:
          type: number
          minimum: 0
          exclusiveMinimum: true
          nullable: true
        photoUrls:
          type: array
          nullable: true
          items:
            type: string
    CreatePartEventRequest:
      type: object
      required: [ eventDateTime, mileageKm, name ]
      properties:
        eventDateTime:
          type: string
          format: date-time
        mileageKm:
          type: integer
          minimum: 0
        name:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        cost:
          type: number
          minimum: 0
          exclusiveMinimum: true
          nullable: true
        photoUrls:
          type: array
          nullable: true
          items:
            type: string
    TimelineEventList:
      type: object
      required: [ events ]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
    TimelineEvent:
      type: object
      description: >-
        Unified timeline event. Fields are populated according to `type`: refuel events use
        liters/fuelType/fuelName/stationName and cost; trip events use
        startMileageKm/endMileageKm/distanceKm/route/durationMinutes; maintenance events use
        name/description/photoUrls and cost.
      required: [ id, type, eventDateTime ]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TimelineEventType'
        title:
          type: string
          nullable: true
        eventDateTime:
          type: string
          format: date-time
        cost:
          type: number
          nullable: true
        mileageKm:
          type: integer
          nullable: true
        liters:
          type: number
          nullable: true
        fuelType:
          allOf:
            - $ref: '#/components/schemas/FuelType'
          nullable: true
        fuelName:
          type: string
          nullable: true
        stationName:
          type: string
          nullable: true
        startMileageKm:
          type: integer
          nullable: true
        endMileageKm:
          type: integer
          nullable: true
        distanceKm:
          type: integer
          nullable: true
        route:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        averageFuelConsumptionLitersPerKm:
          type: number
          nullable: true
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        photoUrls:
          type: array
          nullable: true
          items:
            type: string
    FuelType:
      type: string
      enum: [ GASOLINE, DIESEL, ELECTRIC, HYBRID, OTHER ]
    PartCategory:
      type: string
      enum: [ ENGINE_OIL, OIL_FILTER, AIR_FILTER, BRAKE_PADS, TIMING_BELT, BATTERY, OTHER ]
    PartStatus:
      type: string
      enum: [ OK, ATTENTION, CRITICAL, UNKNOWN, REPLACED ]
    TimelineEventType:
      type: string
      enum: [ TRIP, REFUEL, REPAIR, MAINTENANCE, PART_REPLACEMENT, WARNING ]
    AnalyticsPeriod:
      type: string
      enum: [ MONTH, YEAR, ALL_TIME ]
    AnalyticsOverview:
      type: object
      required: [ period, totalExpenses, currency, expensesByCategory, monthlyExpenses, seasonalExpenses, costPerKilometer, fuel, historyAnalysis, hasData ]
      properties:
        period:
          $ref: '#/components/schemas/AnalyticsPeriod'
        totalExpenses:
          type: number
        currency:
          type: string
          example: RUB
        expensesByCategory:
          type: object
          additionalProperties:
            type: number
        monthlyExpenses:
          type: array
          items:
            $ref: '#/components/schemas/MonthlyExpense'
        seasonalExpenses:
          type: array
          items:
            $ref: '#/components/schemas/SeasonalExpense'
        costPerKilometer:
          $ref: '#/components/schemas/CostPerKilometer'
        fuel:
          $ref: '#/components/schemas/FuelAnalytics'
        historyAnalysis:
          $ref: '#/components/schemas/HistoryAnalysis'
        hasData:
          type: boolean
    MonthlyExpense:
      type: object
      required: [ month, total, breakdownByCategory ]
      properties:
        month:
          type: string
          example: 2026-06
        total:
          type: number
        breakdownByCategory:
          type: object
          additionalProperties:
            type: number
    SeasonalExpense:
      type: object
      required: [ season, total ]
      properties:
        season:
          type: string
          enum: [ WINTER, SPRING, SUMMER, AUTUMN ]
        total:
          type: number
    CostPerKilometer:
      type: object
      required: [ totalKm, totalExpenses, costPerKm ]
      properties:
        totalKm:
          type: integer
        totalExpenses:
          type: number
        costPerKm:
          type: number
    FuelAnalytics:
      type: object
      required: [ totalLiters, averageConsumptionLitersPer100Km ]
      properties:
        totalLiters:
          type: number
        averageConsumptionLitersPer100Km:
          type: number
    HistoryAnalysis:
      type: object
      required: [ eventCount, refuelCount, tripCount, maintenanceCount, partEventCount, totalTripKm, averageTripKm ]
      properties:
        eventCount:
          type: integer
        refuelCount:
          type: integer
        tripCount:
          type: integer
        maintenanceCount:
          type: integer
        partEventCount:
          type: integer
        totalTripKm:
          type: integer
        averageTripKm:
          type: number